[Cleanup] Reduce build size ESPEasySerial inline functions to .cpp
