+++ lib/ESPEasySerial/library.json
@@ -1,6 +1,6 @@
 {
     "name": "ESPeasySerial",
-    "version": "2.0.10",
+    "version": "2.0.11",
     "keywords": [
       "serial", "io", "softwareserial", "hardwareserial"
     ],
