+++ platformio_esp32_envs.ini
@@ -39,7 +39,7 @@ lib_deps                  =
                             adafruit/Adafruit ILI9341@^1.5.10
                             ccronexpr
                             rlogiacco/CircularBuffer
-                            td-er/ESPeasySerial @ 2.0.10
+                            td-er/ESPeasySerial @ 2.0.11
                             td-er/SparkFun MAX1704x Fuel Gauge Arduino Library @ ^1.0.1
                             RAK12019_LTR390_UV_Light
 extra_scripts             = post:tools/pio/post_esp32.py
