+++ platformio_esp82xx_base.ini
@@ -79,7 +79,7 @@ lib_deps                  =
                             adafruit/Adafruit ILI9341@^1.5.10
                             bblanchon/ArduinoJson @ ^6.17.2
                             rlogiacco/CircularBuffer
-                            td-er/ESPeasySerial @ 2.0.10
+                            td-er/ESPeasySerial @ 2.0.11
                             td-er/RABurton ESP8266 Mutex @ ^1.0.2
                             td-er/SparkFun MAX1704x Fuel Gauge Arduino Library @ ^1.0.1
                             RAK12019_LTR390_UV_Light
@@ -142,6 +142,7 @@ platform_packages         = ${regular_platform.platform_packages}
 build_flags               = ${regular_platform.build_flags}
                             -DPLUGIN_BUILD_TESTING
                             -DLIMIT_BUILD_SIZE
+                            -DDISABLE_SC16IS752_Serial
 lib_ignore                = ${regular_platform.lib_ignore}
                             SD(esp8266)
                             SD
@@ -155,6 +156,7 @@ platform_packages         = ${regular_platform_alt_wifi.platform_packages}
 build_flags               = ${regular_platform_alt_wifi.build_flags}
                             -DPLUGIN_BUILD_TESTING
                             -DLIMIT_BUILD_SIZE
+                            -DDISABLE_SC16IS752_Serial
 lib_ignore                = ${regular_platform_alt_wifi.lib_ignore}
                             SD(esp8266)
                             SD
@@ -167,6 +169,7 @@ platform_packages         = ${beta_platform.platform_packages}
 build_flags               = ${beta_platform.build_flags}
                             -DPLUGIN_BUILD_TESTING
                             -DLIMIT_BUILD_SIZE
+                            -DDISABLE_SC16IS752_Serial
 lib_ignore                = ${beta_platform.lib_ignore}
                             SD
                             SDFS
